+++ scratch.js
@@ -7,7 +7,6 @@ const spacetime = require('./src/index')
 // let s = spacetime('2019-04-01T12:15:00.000+04:00')
 // console.log(s.format('iso'))
 
-console.log(spacetime('2019-04-01T12:15:00.000+03:00').format('iso'))
-console.log(spacetime('2019-04-01T12:15:00.000+04:00').format('iso'))
-console.log(spacetime('2019-04-01T12:15:00.000+05:00').format('iso'))
-console.log(spacetime('2019-04-01T12:15:00.000+05:30').format('iso'))
+let s = spacetime.now()
+s.tz = 'UTC'
+console.log(s.timezone())